+++ CHANGELOG.md
@@ -1,6 +1,21 @@
 ## v0.0.7
 
-- Changed Details formate to `FileSize LinesCount : WordsCount : DataCount`
+Added Custom Display Format and some additional statusbar settings
+
+- `data-size-count.statusBar.position` - Set Custom Statusbar position.
+- `data-size-count.statusBar.priority` - The priority of the statusbar. Higher value means the statusbar should be shown more to the left.
+- `data-size-count.statusBar.fileSizeformat` - You can use these variables to custom display format: \${fileSize}.
+- `data-size-count.statusBar.countsformat` - You can use these variables to custom display format: \${linesCount} \${wordsCount} \${dataCount} \${dataCountWithBrackets}.
+- `data-size-count.statusBar.visibility` - Show/Hide File Size or Selection Count details in statusbar.
+- `data-size-count.statusBar.itemSeperator` - Separator between File Size and Selection Counts in statusbar. The default value is 2 blank spaces.
+
+Fixes
+
+- lines count is incorrect when using multi cursor selection - `Fixed`
+
+## v0.0.7
+
+- Changed Details format to `FileSize LinesCount : WordsCount : DataCount`
 
 ## v0.0.6